Move Aircraft.cs to Aircraft directory and reformat file
